What: Refactored function integration tests Why: Improve maintainability
